refactor(tools): extract skill tools to skill_tools.py (audit-B / H2)
cli/tools.py had skill-loading factories (load_skill / unload_skill /
skill_review) and their helpers (_strip_frontmatter / _find_skill_resources /
_get_evolution_hints) physically nested inside SECTION 3 (MEMORY TOOLS).
~500 lines of skill-specific code masquerading as memory tools — the
'wrong category content ran into the section' pattern.

Extract to new loom/platform/cli/skill_tools.py:
- make_load_skill_tool (+ _strip_frontmatter, _find_skill_resources,
  _get_evolution_hints helpers)
- make_unload_skill_tool
- make_skill_review_tool

cli/tools.py changes:
- 4443 → 3947 lines (-496)
- SECTION 3 banner relabeled MEMORY + LEDGER RECALL (matches what's
  actually still there: make_ledger_recall_tool stays — it's not a
  skill tool)
- Pointer comment where the skill block used to be

session.py: imports rerouted to skill_tools.
tests: test_skill_tools.py / test_skill_review_tool.py /
test_skill_evolution_retirement.py import paths updated.
test_architecture_guards.py: skill_tools allowlisted with the same
justification as cli.tools (will move to loom.core.tools under #69).

Verification:
- 1961 passed, 3 failed (pre-existing scope-grant, unrelated)
- skill_tools.py imports clean, all 9 module-level + 6 helper sites
  resolve

Refs #399 (audit-B / H2).
