+++ tests/test_skill_tools.py
@@ -169,7 +169,7 @@ class TestStripFrontmatter:
     """Test frontmatter stripping from skill body."""
 
     def test_strip_removes_yaml(self):
-        from loom.platform.cli.tools import _strip_frontmatter
+        from loom.platform.cli.skill_tools import _strip_frontmatter
 
         body = """---
 name: test
@@ -185,7 +185,7 @@ Instructions here.
         assert "name: test" not in result
 
     def test_no_frontmatter_passthrough(self):
-        from loom.platform.cli.tools import _strip_frontmatter
+        from loom.platform.cli.skill_tools import _strip_frontmatter
 
         body = "# Just markdown\n\nNo frontmatter."
         assert _strip_frontmatter(body) == body
@@ -195,7 +195,7 @@ class TestFindSkillResources:
     """Test skill directory resource discovery."""
 
     def test_finds_directory(self, tmp_path):
-        from loom.platform.cli.tools import _find_skill_resources
+        from loom.platform.cli.skill_tools import _find_skill_resources
 
         # Create a skill directory with resources
         skill_dir = tmp_path / "skills" / "test-skill"
@@ -211,7 +211,7 @@ class TestFindSkillResources:
         assert "scripts\\helper.py" in resources or "scripts/helper.py" in resources
 
     def test_not_found(self, tmp_path):
-        from loom.platform.cli.tools import _find_skill_resources
+        from loom.platform.cli.skill_tools import _find_skill_resources
 
         dir_path, resources = _find_skill_resources(
             "nonexistent", [tmp_path]
@@ -220,7 +220,7 @@ class TestFindSkillResources:
         assert resources == []
 
     def test_underscore_hyphen_variant(self, tmp_path):
-        from loom.platform.cli.tools import _find_skill_resources
+        from loom.platform.cli.skill_tools import _find_skill_resources
 
         # Skill stored with underscore, queried with hyphen
         skill_dir = tmp_path / "skills" / "loom_engineer"
@@ -366,7 +366,7 @@ class TestEvolutionHintsFromSemantic:
 
     @pytest.mark.asyncio
     async def test_reads_real_hints(self):
-        from loom.platform.cli.tools import _get_evolution_hints
+        from loom.platform.cli.skill_tools import _get_evolution_hints
         from loom.core.memory.semantic import SemanticEntry
 
         mock_procedural = AsyncMock()
@@ -387,7 +387,7 @@ class TestEvolutionHintsFromSemantic:
 
     @pytest.mark.asyncio
     async def test_falls_back_to_confidence(self):
-        from loom.platform.cli.tools import _get_evolution_hints
+        from loom.platform.cli.skill_tools import _get_evolution_hints
 
         mock_procedural = AsyncMock()
         mock_semantic = AsyncMock()
@@ -407,7 +407,7 @@ class TestEvolutionHintsFromSemantic:
 
     @pytest.mark.asyncio
     async def test_no_hints_high_confidence(self):
-        from loom.platform.cli.tools import _get_evolution_hints
+        from loom.platform.cli.skill_tools import _get_evolution_hints
 
         mock_procedural = AsyncMock()
         mock_semantic = AsyncMock()
@@ -424,7 +424,7 @@ class TestEvolutionHintsFromSemantic:
     @pytest.mark.asyncio
     async def test_works_without_semantic(self):
         """When semantic is None (e.g. not connected), fallback still works."""
-        from loom.platform.cli.tools import _get_evolution_hints
+        from loom.platform.cli.skill_tools import _get_evolution_hints
 
         mock_procedural = AsyncMock()
         mock_skill = MagicMock()
